added documentation for villages routes
